Set default "from" email address globally
Refactor the code to set the sender email address in one place instead
of multiple places.

This is in preparation for making the "from" address configurable.

Note that `config.action_mailer.default()` doesn't work[1], so we must
use `default_options=`.

Also drop `:charset => "utf-8"`, because the default is already UTF-8
(and that line wasn't even doing anything due to [1]).

[1] https://www.github.com/rails/rails/issues/54700