Write out jquery.facebox.js
As far as I can tell, it's only used exactly once. This is for a modal /
dialog thing that confirms your logged in user when an admin is using
the "su" function to be logged in as another user.

Instead, we can redirect to another page confirm the password then
redirect to wherever su would totally take us.