Write out jquery.facebox.js
It's only used once, for a modal / dialog thing that confirms your
password when an admin is using the "su" function to log in as another
user.

Instead, we can redirect to the dialog as a standalone page (and then
redirect to wherever su would totally take us).

The "Cancel" button was previously broken due to invalid escaping. It
now navigates back to the previous page in the history, or to the
"users" page if there is no history (e.g. because the "su" link was
opened in a new tab).