Add a version constraint on loofah
Loofah 2.21.0 added support for HTML5 (as did a later version of
Nokogiri which loofah is based off). However, we can't upgrade Nokogiri
yet due to a higher ruby version requirement. The nokogiri dependencies
aren't well constrainted for in loofah 2.21.0 so we can just stay below
that version for now, until we can upgrade ruby to 2.5, which is after
we can upgrade Rails to 5.1.

https://github.com/flavorjones/loofah/blob/main/CHANGELOG.md#2210--2023-05-10