Write out the squeel gem
This gem isn't supported on Rails >= 5, and also fully deprecated. I've
replaced all usages of it with the equivalent Rails 4.2 code which isn't
necessarily the most ergonomic nor the most performance, in some cases,
but should be fine for our cases.

I also had to repair one migration that was added back in 2013 and will
never be run again, and could theoretically just be commented out, but
the fix was trivial.