Remove useless line from COCI PDFImporter
The expression used to be on the same line as the "else" (before it
was auto-formatted). It might have been intended as an "elsif", but
the same condition is checked earlier so it's redundant.

The strange code was introduced in commit 01ed6b88 (make COCI importer
more robust, though some parts of the problems are poorly imported,
2014-12-25).