Write out FixturesSpecHelper
Having this helper generate these records before the test suite runs
means they are not included any the standard rspec transactional tests
mechanism. This means that if the suite cannot fully complete for
any reason, such as a syntax error, or if the number of create/destroy
is not perfectly matched in a test then the following test run will
fail.

Using let(:x) and before is best practise, and shouldn't have a
significant effect on test runtimes.