Remove unused actions on ContestRelationsController
Most of these actions were rendered unused in 2da020ef59c5, #destroy was
restored in 112b236cf63e and #update_year_level added in 00c9636d1364.
The rest of these actions seem to be dead code, so lets remove them! Git
history can always bring them back as needed, but none of the code is
particularly interesting so it'd be easy to rewrite, anyway.

I've also added specs to cover off the major functions of
contest_relations_controller, including some permission tests, and
ensured that there is at least a logged in user to slightly mitigate a
leak of information in the response differing between not_found and
unauthorised on ContestRelation objects.