Remove most XML responses
Almost none of these were being used usefully (as checked by a quick
grep of the logs on the server) except for SubmissionController#show,
which I have left in place for now. I also removed all comments refering
to routes, as a great number of them are incorrect and we discussed
briefly on discord and decided to remove. (use `bundle exec rake routes`
instead. Additionally, a fixed a couple of cases of record#destory being
called instead of record#destroy!. You pretty much always want to use
destroy! - the former version has no errors or warnings if the delete
fails for any reason, which is rarely what you want.