Expose the submission#show action as a XML when requested
Admin users can add `?debug` to the URL to get the XML dump of the
specific submission, which can be useful for debugging certain issues
with judging etc.

This replaces using `.xml` to request a different format, and means we
don't need to be careful about restricting formats based on the logged
in user - instead, we can decide what we want to expose irrespective of
format, and rely on the permission model instead.

Longer term - it might be a better idea to add a "debug" tab to the
submission view for admin users, but this will do for now.