Refactor db/seeds to allow for easy development seeds
In order to try out different things and qa changes locally it can be
helpful to have various data in your development database. This just
makes it a little easier and safer to load that data into the dev
database repeatedly.

Also fix up how languages.yml are loaded by switching to YAML.safe_load
and dropping the symbol keys in favour of using HashWithIndifferentAccess.