Run rails:app update and apply changes as required
The environment configuration in config/environments/*.rb hasn't been
updated to the rails defaults with out modifications, as there are a lot
of individual changes and this is best left for future discussions and
changes.

Especially in the context of another ~8 rails versions to
upgrade to get anywhere near the latest, there will be a lot of churn in
those files and minimising that where possible is the goal here.