Configure the hybrid cookie session store
Rails 4.1 introduced the new JSON format, and it becomes default in
Rails 5.0. We want to be using hybrid in production for a period of time
in order to migrate everyone to JSON before letting JSON be the default.

Rails 7.0 changes the default to :json, but as long as we leave this
initialiser in place we will continue to use hybrid. There are very very
minor security concerns to allowing hybrid (and therefore marshall) but
there are other tradeoffs to using :json, too.

https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#cookies-serializer
https://guides.rubyonrails.org/configuring.html#config-action-dispatch-cookies-serializer