Implement expanding row items in terms of aria-expanded & css
Instead of relying on jQuery doing fancy things with the hide and show,
we can just use a css-transition. Unfortunately, css-transitions don't
really support animating max-height very nicely between 0 and auto, so
instead we animate between a 0 and a number larger than our biggest
problem set.

We use a cubic-bezier() animation so we can cheekily invert the
animation curve between the two states, leading to a nicer closing
experience when the content itself is larger than the max-height.

Using aria-expanded gives us a very slightly nicer experience on
screen-readers