Remove JS and CSS asset compressors
Rails 8 has moved away from asset compression[1].

`uglifier` (JS compressor) requires a JavaScript runtime (which we
previously provided using mini_racer). It's a large install and raises
security concerns because it uses an outdated version of V8 (it's only
meant to run as part of asset precompilation, not in the running web
server; but still).

`yui` requires the Java Runtime Environment.

We follow the Rails 8 rationale and remove the asset compression (even
though we are still on a much older version of Rails).

[1] https://guides.rubyonrails.org/v8.0/asset_pipeline.html