Remove config/secrets.yml
This file is deleted later in the Rails upgrade series, so it feels a
bit silly to add it now. Additionally, having it be present with a
different value to the values we set for test and development elsewhere
causes annoying and silly issues when switching branches while we
develop and review this branch.