Remove usage of alias_method_chain
alias_method_chain is deprecated in Rails 5.0 and removed in Rails 5.1,
so we need to stop using it.

This allowed us to modify the behaviour of `ContestRelation#contest=`
while still providing access to the original method through
`contest_without_update=`. But since we weren't using that method
anyway, we can just directly override `contest=` instead.

Co-authored-by: Jonathan Khoo <jkhoo98@gmail.com>