Add tests for (and fix) the after_save behaviour for contest
This after_save hook is supposed to be handling the cases where the
contest is changed, but there are all sorts of things going on that
means it's being overwritten or ignored or just not called at all.

I'd like to write out the weird stuff around contest_relation#finish_at,
so this is step 1 of that.

before_save is changed to after_save, as that's really what it should
have been in the first place. The only reason it works now is that the
calculation of finish_at is done inside `Contest` instead of
`ContestRelation` which is 'a bit odd', but fine.