Refactor contest#after_save
The prior version of this code is a bit hard to follow as also has one
hook doing two responsibilities. This is just a minor factor to
hopefully clear it all up a bit and move some of the responsibilities
around.

The long term goal is to let ContestRelation calculate it's own
finish_at from its parent Contest, then changes to Contest just ask
ContestRelation to recalculate, like every other way a contest_relation
might need to be updated.

We _could_ use touch: true on the has_many so that any saved changed to
contest causes an update of contest_relation, then have the
set_finish_at be a before_validation callback on ContestRelation, but I
always feel a little weird about the implictness of that.