Correctly manage qless during the deploy lifecycle (#345)
We want qless stopped before we run migrations, as a minor safety
feature and we can afford to have a minute or so of downtime of that job
system during deploys in return for not having to be quite so agressive
about migration safety that could affect running jobs.

We boot it back up again at the end.